giving the user more informed error messages when registering
